% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorSeason.R
\docType{data}
\name{colorSeason.palet.info}
\alias{colorSeason.palet.info}
\title{season color palettes infomation}
\format{A data frame}
\usage{
colorSeason.palet.info
}
\description{
my season color palettes's info. 2018 R HW
}
\examples{
devtools::install_github("shb0328/RColorSeason")
library(colorseason)
colorSeason.palet.info

}
\keyword{datasets}
