% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ColorSeason.R
\name{display.colorSeason.palet}
\alias{display.colorSeason.palet}
\title{display season color palettes}
\usage{
display.colorSeason.palet(name)
}
\arguments{
\item{name}{Spring Summer Fall Winter}
}
\description{
function for my season color palettes. 2018 R HW
}
\examples{
devtools::install_github("shb0328/RColorSeason")
library(colorseason)
display.colorSeason.palet("Winter")
devAskNewPage(ask=TRUE)

}
